fix: move highlight documents to luaCAT comments
